Nutrient number can be double

+++ app/src/main/java/com/android/project/lightweight/api/model/FoodNutrient.kt
@@ -10,5 +10,5 @@ data class FoodNutrient(
     val unitName : String,
     @Json(name = "value")
     val amount : Double, // amount in 100 "unitName" e.g.: 100 g, mg etc.
-    val nutrientNumber : Int
+    val nutrientNumber : Double
 ) : Parcelable
